Convert OptionsPanel to TS
